extract read module
