+++ with-application-events-immutable/src/main/java/io/dddbyexamples/cqrs/read/WithdrawalFinder.java
@@ -1,0 +1,23 @@
+package io.dddbyexamples.cqrs.read;
+
+import lombok.AllArgsConstructor;
+import org.springframework.jdbc.core.BeanPropertyRowMapper;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+
+import java.util.List;
+import java.util.UUID;
+
+@Service
+@AllArgsConstructor
+public class WithdrawalFinder {
+
+    private final JdbcTemplate jdbcTemplate;
+
+
+    public List<WithdrawalReadModel> loadWithdrawalsFor(@PathVariable UUID cardId) {
+        return jdbcTemplate.query("SELECT * FROM WITHDRAWAL WHERE CARD_ID = ?", new Object[]{cardId},
+                new BeanPropertyRowMapper<>(WithdrawalReadModel.class));
+    }
+}
