+++ with-application-events-immutable/src/main/java/io/dddbyexamples/cqrs/read/WithdrawalReadModel.java
@@ -1,4 +1,4 @@
-package io.dddbyexamples.cqrs.ui;
+package io.dddbyexamples.cqrs.read;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -10,7 +10,7 @@ import java.util.UUID;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class WithdrawalDto {
+public class WithdrawalReadModel {
     private UUID cardId;
     private BigDecimal amount;
