+++ with-application-events-immutable/src/main/java/io/dddbyexamples/cqrs/ui/WithdrawalsController.java
@@ -1,25 +1,23 @@
 package io.dddbyexamples.cqrs.ui;
 
 import io.dddbyexamples.cqrs.application.WithdrawalProcess;
+import io.dddbyexamples.cqrs.read.WithdrawalFinder;
+import io.dddbyexamples.cqrs.read.WithdrawalReadModel;
+import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.*;
 
 import javax.websocket.server.PathParam;
 import java.util.List;
 import java.util.UUID;
 
+
+@AllArgsConstructor
 @RestController("/withdrawals")
 class WithdrawalsController {
 
-    private final JdbcTemplate jdbcTemplate;
     private final WithdrawalProcess withdrawalProcess;
-
-    WithdrawalsController(JdbcTemplate jdbcTemplate, WithdrawalProcess withdrawalsProcess) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.withdrawalProcess = withdrawalsProcess;
-    }
+    private final WithdrawalFinder finder;
 
     @PostMapping
     ResponseEntity withdraw(@RequestBody WithdrawalCommand withdrawalCommand) {
@@ -28,12 +26,8 @@ class WithdrawalsController {
     }
 
     @GetMapping
-    ResponseEntity<List<WithdrawalDto>> withdrawals(@PathParam("cardId") String cardId) {
-        return ResponseEntity.ok().body(loadWithdrawalsFor(UUID.fromString(cardId)));
+    ResponseEntity<List<WithdrawalReadModel>> withdrawals(@PathParam("cardId") String cardId) {
+        return ResponseEntity.ok().body(finder.loadWithdrawalsFor(UUID.fromString(cardId)));
     }
 
-    private List<WithdrawalDto> loadWithdrawalsFor(@PathVariable UUID cardId) {
-        return jdbcTemplate.query("SELECT * FROM WITHDRAWAL WHERE CARD_ID = ?", new Object[]{cardId},
-                new BeanPropertyRowMapper<>(WithdrawalDto.class));
-    }
 }
